Modified initial code to assign bins via trial type
Modified the codes to assign bins to each of the two congruency trial type based on the value of the CueFixation_mean variable, this is done for each individual participant.

The code generates 3 csv files; the FET_TrialLevelDataACC.PTACC_bins.csv is the trial by trial data with the bins assigned to each trial for each participants, FET_summary.csv and FET_varify.csv are files for checking if the code functioned properly. The former provides information on number of trials, trials in each bin, and number of remainder for the two congruency types for each participant . The latter counts the number of trials in each bin for the two congruency type for each participant, this allow us to see if the number of trials assigned to each bin is accurate.